feat(module-resolver): Support for module resolver
feat(asset/folder): Support for external assets folder

feat(types/expo): Added @types/expo

refact(typings): Migrate for @types/expo and move screen files to folder
