+++ .babelrc
@@ -1,8 +1,24 @@
 {
-  "presets": ["babel-preset-expo"],
+  "presets": [
+    "babel-preset-expo"
+  ],
   "env": {
     "development": {
-      "plugins": ["transform-react-jsx-source"]
+      "plugins": [
+        "transform-react-jsx-source",
+        [
+          "module-resolver",
+          {
+            "cwd": "babelrc",
+            "root": [
+              "./src"
+            ],
+            "alias": {
+              "assets": "./assets"
+            }
+          }
+        ]
+      ]
     }
   }
 }
