+++ src/constants/ConstantsScreen.tsx
@@ -1,60 +1,59 @@
 import * as React from 'react'
 import { Component } from 'react'
 import { Constants } from 'expo'
-import { Text } from 'react-native'
-import { View } from 'react-native'
+import { Text, View } from 'react-native'
 import { NavigationScreenProps } from 'react-navigation'
 
 interface State {
-  webViewUserAgent: string | undefined
+    webViewUserAgent: string | undefined
 }
 
 export class ConstantsScreen extends Component<NavigationScreenProps, State> {
-  constructor(props: NavigationScreenProps, context?: any) {
-    super(props, context)
-
-    this.state = {
-      webViewUserAgent: undefined
+    public static navigationOptions = {
+        title: 'Constants',
     }
 
-    this.updateWebViewUserAgent()
-  }
+    constructor(props: NavigationScreenProps, context?: any) {
+        super(props, context)
 
-  public static navigationOptions = {
-    title: 'Constants'
-  }
+        this.state = {
+            webViewUserAgent: undefined,
+        }
 
-  public render() {
-    const navigate = this.props.navigation.navigate
+        this.updateWebViewUserAgent()
+    }
 
-    return (
-      <View
-        style={{
-          flex: 1,
-          padding: 10
-        }}
-      >
-        <Text>appOvernership: "{Constants.appOwnership}"</Text>
-        <Text>deviceId: "{Constants.deviceId}"</Text>
-        <Text>deviceName: "{Constants.deviceName}"</Text>
-        <Text>deviceYearClass: {Constants.deviceYearClass}</Text>
-        <Text>expoVersion: "{Constants.expoVersion}"</Text>
-        <Text>isDevice: {Constants.isDevice}</Text>
-        <Text>linkingUri: "{Constants.linkingUri}"</Text>
-        <Text onPress={() => navigate('Manifest')}>manifest: (Click to see)</Text>
-        <Text onPress={() => navigate('Platform')}>platform: (Click to see)</Text>
-        <Text>sessionId: "{Constants.sessionId}"</Text>
-        <Text>statusBarHeight: {Constants.statusBarHeight}</Text>
-        <Text onPress={() => navigate('SystemFonts')}>systemFonts: (Click to see)</Text>
-        <Text>webViewUserAgent: "{this.state.webViewUserAgent}"</Text>
-      </View>
-    )
-  }
+    public render() {
+        const navigate = this.props.navigation.navigate
 
-  private async updateWebViewUserAgent() {
-    const userAgent = await Constants.getWebViewUserAgentAsync()
-    this.setState({
-      webViewUserAgent: userAgent
-    })
-  }
+        return (
+            <View
+                style={{
+                    flex: 1,
+                    padding: 10,
+                }}
+            >
+                <Text>appOvernership: "{Constants.appOwnership}"</Text>
+                <Text>deviceId: "{Constants.deviceId}"</Text>
+                <Text>deviceName: "{Constants.deviceName}"</Text>
+                <Text>deviceYearClass: {Constants.deviceYearClass}</Text>
+                <Text>expoVersion: "{Constants.expoVersion}"</Text>
+                <Text>isDevice: {Constants.isDevice}</Text>
+                <Text>linkingUri: "{Constants.linkingUri}"</Text>
+                <Text onPress={() => navigate('Manifest')}>manifest: (Click to see)</Text>
+                <Text onPress={() => navigate('Platform')}>platform: (Click to see)</Text>
+                <Text>sessionId: "{Constants.sessionId}"</Text>
+                <Text>statusBarHeight: {Constants.statusBarHeight}</Text>
+                <Text onPress={() => navigate('SystemFonts')}>systemFonts: (Click to see)</Text>
+                <Text>webViewUserAgent: "{this.state.webViewUserAgent}"</Text>
+            </View>
+        )
+    }
+
+    private async updateWebViewUserAgent() {
+        const userAgent = await Constants.getWebViewUserAgentAsync()
+        this.setState({
+            webViewUserAgent: userAgent,
+        })
+    }
 }
