+++ src/constants/ManifestScreen.tsx
@@ -1,19 +1,18 @@
 import * as React from 'react'
-import { Text } from 'react-native'
 import { Component } from 'react'
+import { ScrollView, Text } from 'react-native'
 import { Constants } from 'expo'
-import { ScrollView } from 'react-native'
 
 export class ManifestScreen extends Component {
-  public static navigationOptions = {
-    title: 'manifest'
-  }
+    public static navigationOptions = {
+        title: 'manifest',
+    }
 
-  public render() {
-    return (
-      <ScrollView>
-        <Text>{JSON.stringify(Constants.manifest, undefined, 2)}</Text>
-      </ScrollView>
-    )
-  }
+    public render() {
+        return (
+            <ScrollView>
+                <Text>{JSON.stringify(Constants.manifest, undefined, 2)}</Text>
+            </ScrollView>
+        )
+    }
 }
