+++ tslint.json
@@ -12,6 +12,8 @@
     ]
   },
   "rules": {
+    "no-implicit-dependencies": false,
+    "no-submodule-imports": false,
     "align": [
       false
     ],
